A start for the final
I played around with creating the matrix and scanning for open spaces.
I made three functions representing a 1x1, 2x2, and 3x3 within the
matrix. They check for an intial open spot and checks that there’s
enough space for the dimension, if so, it changes that area to the
desired image, in this case number representative.